+++ lib/src/models/compression_preset.dart
@@ -14,4 +14,3 @@ enum CompressionPreset {
     );
   }
 }
-